refactor: fix circular handler shift in models route, fix types.ts HttpProxyConfig and any usage
