+++ src/app/api/providers/[id]/models/handlers/azure_openai.ts
@@ -81,7 +81,7 @@ import { getSyncedAvailableModels } from "@/lib/db/models";
 import { fetchCursorAgentModels } from "@/lib/providerModels/cursorAgent";
 
 import { ModelsRequestContext } from "../types.ts";
-import { getProviderBaseUrl } from "../utils.ts";
+import { getProviderBaseUrl, normalizeAzureOpenAIBaseUrl, getAzureOpenAIApiVersion, normalizeOpenAiLikeModelsResponse } from "../utils.ts";
 import { GET } from "../route.ts";
 
 export async function handleAzureOpenaiModels(ctx: ModelsRequestContext): Promise<any> {
@@ -94,11 +94,6 @@ export async function handleAzureOpenaiModels(ctx: ModelsRequestContext): Promis
 
       const token = accessToken || apiKey;
       if (!token) {
-        const fallback = buildDiscoveryFallbackResponse({
-          cacheWarning: "No token configured — using cached catalog",
-          localWarning: "No token configured — using local catalog",
-        });
-        if (fallback) return fallback;
         return NextResponse.json(
           {
             error:
@@ -108,57 +103,64 @@ export async function handleAzureOpenaiModels(ctx: ModelsRequestContext): Promis
         );
       }
 
-      const baseUrl =
-        getProviderBaseUrl(connection.providerSpecificData) || AZURE_AI_DEFAULT_BASE_URL;
-      const modelsUrl = buildAzureAiModelsUrl(baseUrl);
-
-      let response: Response;
-      try {
-        response = await safeOutboundFetch(modelsUrl, {
-          ...SAFE_OUTBOUND_FETCH_PRESETS.modelsDiscovery,
-          guard: getProviderOutboundGuard(),
-          proxyConfig: proxy,
-          method: "GET",
-          headers: {
-            "Content-Type": "application/json",
-            "api-key": token,
-          },
-        });
-      } catch (error) {
-        const fallback = buildDiscoveryErrorFallbackResponse(error, {
-          cacheWarning: "Azure AI models API unavailable — using cached catalog",
-          localWarning: "Azure AI models API unavailable — using local catalog",
-        });
-        if (fallback) return fallback;
-        throw error;
-      }
-
-      if (!response.ok) {
-        const fallback = buildDiscoveryFallbackResponse({
-          cacheWarning: `Models probe failed (${response.status}) — using cached catalog`,
-          localWarning: `Models probe failed (${response.status}) — using local catalog`,
-        });
-        if (fallback) return fallback;
+      const rawBaseUrl = getProviderBaseUrl(connection.providerSpecificData);
+      if (!rawBaseUrl) {
         return NextResponse.json(
-          { error: `Failed to fetch models: ${response.status}` },
-          { status: response.status }
+          { error: "No Azure OpenAI resource endpoint configured" },
+          { status: 400 }
         );
       }
 
-      const data = await response.json();
-      const models = (data.data || data.models || []).map((model: Record<string, unknown>) => ({
-        id:
-          (typeof model.id === "string" && model.id) ||
-          (typeof model.name === "string" && model.name) ||
-          "",
-        name:
-          (typeof model.display_name === "string" && model.display_name) ||
-          (typeof model.name === "string" && model.name) ||
-          (typeof model.id === "string" && model.id) ||
-          "",
-        owned_by: "azure-ai",
-      }));
+      const baseUrl = normalizeAzureOpenAIBaseUrl(rawBaseUrl);
+      const apiVersion = encodeURIComponent(
+        getAzureOpenAIApiVersion(connection.providerSpecificData)
+      );
+      const discoveryUrls = [
+        `${baseUrl}/openai/deployments?api-version=${apiVersion}`,
+        `${baseUrl}/openai/models?api-version=${apiVersion}`,
+      ];
 
-      return buildApiDiscoveryResponse(models.filter((model) => model.id));
+      let lastStatus = 0;
+      for (const modelsUrl of discoveryUrls) {
+        let response: Response;
+        try {
+          response = await safeOutboundFetch(modelsUrl, {
+            ...SAFE_OUTBOUND_FETCH_PRESETS.modelsDiscovery,
+            guard: getProviderOutboundGuard(),
+            proxyConfig: proxy,
+            method: "GET",
+            headers: {
+              "Content-Type": "application/json",
+              "api-key": token,
+            },
+          });
+        } catch (error) {
+          const fallback = buildDiscoveryErrorFallbackResponse(error, {
+            cacheWarning: "Azure OpenAI models API unavailable — using cached catalog",
+            localWarning: "Azure OpenAI models API unavailable — using local catalog",
+          });
+          if (fallback) return fallback;
+          throw error;
+        }
+
+        if (response.ok) {
+          return buildApiDiscoveryResponse(
+            normalizeOpenAiLikeModelsResponse(await response.json(), "azure-openai")
+          );
+        }
+
+        lastStatus = response.status;
+        if (response.status === 401 || response.status === 403) break;
+      }
+
+      const fallback = buildDiscoveryFallbackResponse({
+        cacheWarning: `Azure OpenAI models probe failed (${lastStatus}) — using cached catalog`,
+        localWarning: `Azure OpenAI models probe failed (${lastStatus}) — using local catalog`,
+      });
+      if (fallback) return fallback;
+      return NextResponse.json(
+        { error: `Failed to fetch models: ${lastStatus || "unknown"}` },
+        { status: lastStatus || 502 }
+      );
   return null;
 }
