+++ src/app/api/providers/[id]/models/handlers/inner_ai.ts
@@ -91,18 +91,100 @@ export async function handleInnerAiModels(ctx: ModelsRequestContext): Promise<an
       if (autoFetchDisabledResponse) return autoFetchDisabledResponse;
 
       try {
-        const models = await fetchCursorAgentModels();
+        // Parse "TOKEN EMAIL" credential format
+        const raw = apiKey.trim();
+        const eqIdx = raw.indexOf("=");
+        const stripped = eqIdx > 0 && !raw.startsWith("eyJ") ? raw.slice(eqIdx + 1).trim() : raw;
+        const lastSpace = stripped.lastIndexOf(" ");
+        let innerAiToken = stripped;
+        let innerAiEmail = "";
+        if (lastSpace > 0) {
+          const possibleEmail = stripped.slice(lastSpace + 1).trim();
+          if (possibleEmail.includes("@")) {
+            innerAiToken = stripped.slice(0, lastSpace).trim();
+            innerAiEmail = possibleEmail;
+          }
+        }
+
+        // Decode device_id from JWT payload
+        let innerAiDeviceId = "";
+        try {
+          const parts = innerAiToken.split(".");
+          if (parts.length >= 2) {
+            const b64 = parts[1].replace(/-/g, "+").replace(/_/g, "/");
+            const payload = JSON.parse(Buffer.from(b64, "base64").toString("utf8"));
+            innerAiDeviceId = String(
+              payload?.device_id ??
+                payload?.deviceId ??
+                payload?.["device-id"] ??
+                payload?.did ??
+                ""
+            ).trim();
+          }
+        } catch {
+          /* ignore */
+        }
+
+        const innerAiHeaders: Record<string, string> = {
+          "USER-TOKEN": innerAiToken,
+          "User-Agent":
+            "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/131.0.0.0 Safari/537.36",
+          Origin: "https://app.innerai.com",
+          Referer: "https://app.innerai.com/",
+        };
+        if (innerAiEmail) innerAiHeaders["USER-EMAIL"] = innerAiEmail;
+        if (innerAiDeviceId) innerAiHeaders["DEVICE-ID"] = innerAiDeviceId;
+
+        const modelsResp = await safeOutboundFetch(
+          "https://platformapi.innerai.com/api/v1/ai_models",
+          { headers: innerAiHeaders },
+          getProviderOutboundGuard(provider)
+        );
+        if (!modelsResp.ok) {
+          throw new Error(`Inner.ai models API returned HTTP ${modelsResp.status}`);
+        }
+
+        const modelsBody = await modelsResp.json().catch(() => null);
+        const rawModels: Array<Record<string, unknown>> = Array.isArray(modelsBody?.ai_models)
+          ? modelsBody.ai_models
+          : Array.isArray(modelsBody)
+            ? modelsBody
+            : [];
+
+        // Filter: enabled, available, text/chat category only.
+        // Use ai_model_categories[].unique_identifier === "text" when available;
+        // fall back to llm_model name heuristic for models without categories.
+        const nonTextPattern =
+          /image|video|audio|img|vid|sound|music|voice|tts|stt|track|clip|avatar|cartoon|flux|stable.diff|recraft|ideogram|leonardo|magnific|bria|seedream|luma|kling|pika|veo|wan-|heygen|did-|vidu|pixverse|sora-|gen-[0-9]|playground|gemini-fal|gamma|lyria|clothes|whisper/i;
+        const textModels = rawModels.filter((m) => {
+          if (m.enable === false || m.unavailable_api) return false;
+          if (typeof m.llm_model !== "string") return false;
+          const cats = Array.isArray(m.ai_model_categories) ? m.ai_model_categories : null;
+          if (cats && cats.length > 0) {
+            return cats.some(
+              (c: Record<string, unknown>) =>
+                String(c.unique_identifier ?? c.name ?? "").toLowerCase() === "text"
+            );
+          }
+          // No categories field — fall back to name heuristic
+          return !nonTextPattern.test(m.llm_model as string);
+        });
+
+        const models = textModels.map((m) => ({
+          id: String(m.llm_model),
+          name: String(m.name || m.llm_model),
+        }));
+
         return buildApiDiscoveryResponse(models);
       } catch (err) {
         const message = err instanceof Error ? err.message : String(err);
-        console.log("[models] cursor-agent fetch failed:", message);
         const fallback = buildDiscoveryFallbackResponse({
-          cacheWarning: `cursor-agent unavailable (${message}) — using cached catalog`,
-          localWarning: `cursor-agent unavailable (${message}) — using local catalog`,
+          cacheWarning: `Inner.ai models unavailable (${message}) — using cached catalog`,
+          localWarning: `Inner.ai models unavailable (${message}) — using local catalog`,
         });
         if (fallback) return fallback;
         return NextResponse.json(
-          { error: `Failed to fetch Cursor models: ${message}` },
+          { error: `Failed to fetch Inner.ai models: ${message}` },
           { status: 502 }
         );
       }
