+++ src/app/api/providers/[id]/models/types.ts
@@ -1,87 +1,12 @@
 import { NextResponse } from "next/server";
-import type { HttpProxyConfig } from "@/shared/utils/http-proxy";
-import {
-  isClaudeCodeCompatibleProvider,
-  isAnthropicCompatibleProvider,
-  isOpenAICompatibleProvider,
-  isSelfHostedChatProvider,
-  NOAUTH_PROVIDERS,
-} from "@/shared/constants/providers";
-import { getRegistryEntry } from "@omniroute/open-sse/config/providerRegistry.ts";
-import { getModelsByProviderId } from "@/shared/constants/models";
-import { getStaticModelsForProvider, type LocalCatalogModel } from "@/lib/providers/staticModels";
-import {
-  getProviderConnectionById,
-  getModelIsHidden,
-  resolveProxyForProvider,
-} from "@/lib/localDb";
-import {
-  SAFE_OUTBOUND_FETCH_PRESETS,
-  SafeOutboundFetchError,
-  getSafeOutboundFetchErrorStatus,
-  safeOutboundFetch,
-} from "@/shared/network/safeOutboundFetch";
-import { getProviderOutboundGuard } from "@/shared/network/outboundUrlGuard";
-import { sanitizeErrorMessage } from "@omniroute/open-sse/utils/error";
-import { getStaticQoderModels } from "@omniroute/open-sse/services/qoderCli.ts";
-import { fetchGitHubCopilotModels } from "@omniroute/open-sse/services/githubCopilotModels.ts";
-import { getAntigravityHeaders } from "@omniroute/open-sse/services/antigravityHeaders.ts";
-import { ensureAntigravityProjectAssigned } from "@omniroute/open-sse/services/antigravityProjectBootstrap.ts";
-import {
-  getAntigravityModelsDiscoveryUrls,
-  getAntigravityFetchAvailableModelsUrls,
-} from "@omniroute/open-sse/config/antigravityUpstream.ts";
-import {
-  buildGlmCodingHeaders,
-  buildGlmModelsUrl,
-} from "@omniroute/open-sse/config/glmProvider.ts";
-import { getImageProvider } from "@omniroute/open-sse/config/imageRegistry.ts";
-import { getVideoProvider } from "@omniroute/open-sse/config/videoRegistry.ts";
-import { resolveAntigravityVersion } from "@omniroute/open-sse/services/antigravityVersion.ts";
-import {
-  discoverBedrockNativeModels,
-  isBedrockNativeApiError,
-} from "@omniroute/open-sse/services/bedrock.ts";
-import {
-  AZURE_AI_DEFAULT_BASE_URL,
-  buildAzureAiModelsUrl,
-} from "@omniroute/open-sse/config/azureAi.ts";
-import {
-  DATAROBOT_DEFAULT_BASE_URL,
-  buildDataRobotCatalogUrl,
-  isDataRobotDeploymentUrl,
-} from "@omniroute/open-sse/config/datarobot.ts";
-import { OCI_DEFAULT_BASE_URL, buildOciModelsUrl } from "@omniroute/open-sse/config/oci.ts";
-import {
-  SAP_DEFAULT_BASE_URL,
-  buildSapModelsUrl,
-  getSapResourceGroup,
-} from "@omniroute/open-sse/config/sap.ts";
-import {
-  WATSONX_DEFAULT_BASE_URL,
-  buildWatsonxModelsUrl,
-} from "@omniroute/open-sse/config/watsonx.ts";
-import {
-  getClientVisibleAntigravityModelName,
-  isUserCallableAntigravityModelId,
-  toClientAntigravityModelId,
-} from "@omniroute/open-sse/config/antigravityModelAliases.ts";
-import { normalizeAntigravityClientProfile } from "@/shared/constants/antigravityClientProfile";
-import { getEmbeddingProvider } from "@omniroute/open-sse/config/embeddingRegistry.ts";
-import { getRerankProvider } from "@omniroute/open-sse/config/rerankRegistry.ts";
-import {
-  getSpeechProvider,
-  getTranscriptionProvider,
-} from "@omniroute/open-sse/config/audioRegistry.ts";
-import {
-  getCachedDiscoveredModels,
-  isAutoFetchModelsEnabled,
-  persistDiscoveredModels,
-} from "@/lib/providerModels/modelDiscovery";
-import { getSyncedAvailableModels } from "@/lib/db/models";
-import { fetchCursorAgentModels } from "@/lib/providerModels/cursorAgent";
 
-import { GET } from "./route.ts";
+export type HttpProxyConfig = {
+  host?: string;
+  port?: number;
+  protocol?: string;
+  auth?: { username: string; password: string };
+  [key: string]: unknown;
+};
 
 export type JsonRecord = Record<string, unknown>;
 
@@ -93,24 +18,33 @@ export type ProviderModelsConfigEntry = {
   authPrefix?: string;
   authQuery?: string;
   body?: unknown;
-  parseResponse: (data: any) => any;
+  parseResponse: (data: unknown) => unknown;
 };
-
-
 
 export interface ModelsRequestContext {
   provider: string;
   connectionId: string;
-  connection: any;
+  connection: {
+    providerSpecificData?: unknown;
+    id?: string;
+    apiKey?: string;
+    accessToken?: string;
+    authType?: string;
+    [key: string]: unknown;
+  };
   apiKey: string;
   accessToken: string;
   proxy: HttpProxyConfig | undefined;
   id: string;
   maybeReturnCachedDiscovery: () => NextResponse | null;
   maybeReturnAutoFetchDisabled: () => NextResponse | null;
-  buildDiscoveryFallbackResponse: (args: { cacheWarning?: string; localWarning?: string; error?: any }) => NextResponse | null;
-  buildDiscoveryErrorFallbackResponse: (error: any, message: string) => NextResponse | null;
-  buildApiDiscoveryResponse: (models: any[], warning?: string) => Promise<NextResponse>;
-  buildResponse: (payload: any, statusConfig?: ResponseInit) => NextResponse;
+  buildDiscoveryFallbackResponse: (args: {
+    cacheWarning?: string;
+    localWarning?: string;
+    error?: unknown;
+  }) => NextResponse | null;
+  buildDiscoveryErrorFallbackResponse: (error: unknown, message: string) => NextResponse | null;
+  buildApiDiscoveryResponse: (models: unknown[], warning?: string) => Promise<NextResponse>;
+  buildResponse: (payload: unknown, statusConfig?: ResponseInit) => NextResponse;
   buildLocalCatalogResponse: (warning?: string) => NextResponse;
 }
